+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Horizontal overflow style to `<StackedAreaChart />`, which were unintentionally removed
 - Use `<ChartContainer />` in `<SparkBarChart />`.
+- Use `<ChartContainer />` in `<SparkLineChart />`.
 
 ### Changed